pyproject.toml
[tool.poetry]
name = "bumppy"
version = "0.1.0"
description = ""
authors = ["Matej Penciak <matej.penciak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gitpython = "^3.1.31"
pygithub = "^1.58.1"
tomli = "^2.0.1"
tomli-w = "^1.0.0"
networkx = "^3.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
